requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-binance>=1.0.0
google-genai>=0.8.0
python-dotenv>=1.0.0
colorama>=0.4.0
pytz>=2023.0
